Create readme
First Task. 
Implement a method index(URLs), which – given a list of input URLs in the array– –  downloads each web page 
–  strips tags in the resulting HTML
–  normalizes the text data (e.g., using string.lower() and the module stemming.porter2)
–  computes and stores a sparse bag-of-words representation of each page (e.g., using python dictionaries)

run your method to index a list of URLs

Implement a method search(query), which – given a sequence of query words – outputs a ranked list of web pages in the index based on the cosine distance measure between the bag-of-words representations of queries and web pages.